refactoring: validateRoles...() -> trait
