+++ app/WebModule/Forms/RolesFormFactory.php
@@ -30,6 +30,7 @@ use Throwable;
 class RolesFormFactory
 {
     use Nette\SmartObject;
+    use RoleFormTrait;
 
     /**
      * Přihlášený uživatel.
@@ -163,77 +164,5 @@ class RolesFormFactory
         } elseif ($form->isSubmitted() === $form['cancelRegistration']) {
             $this->applicationService->cancelRegistration($this->user, ApplicationState::CANCELED, $this->user);
         }
-    }
-
-    /**
-     * Ověří kapacitu rolí.
-     */
-    public function validateRolesCapacities(MultiSelectBox $field): bool
-    {
-        $selectedRoles = $this->roleRepository->findRolesByIds($field->getValue());
-
-        return $this->validators->validateRolesCapacities($selectedRoles, $this->user);
-    }
-
-    /**
-     * Ověří kompatibilitu rolí.
-     *
-     * @param Role[] $args
-     */
-    public function validateRolesIncompatible(MultiSelectBox $field, array $args): bool
-    {
-        $selectedRoles = $this->roleRepository->findRolesByIds($field->getValue());
-        $testRole      = $args[0];
-
-        return $this->validators->validateRolesIncompatible($selectedRoles, $testRole);
-    }
-
-    /**
-     * Ověří výběr vyžadovaných rolí.
-     *
-     * @param Role[] $args
-     */
-    public function validateRolesRequired(MultiSelectBox $field, array $args): bool
-    {
-        $selectedRoles = $this->roleRepository->findRolesByIds($field->getValue());
-        $testRole      = $args[0];
-
-        return $this->validators->validateRolesRequired($selectedRoles, $testRole);
-    }
-
-    /**
-     * Ověří registrovatelnost rolí.
-     */
-    public function validateRolesRegisterable(MultiSelectBox $field): bool
-    {
-        $selectedRoles = $this->roleRepository->findRolesByIds($field->getValue());
-
-        return $this->validators->validateRolesRegisterable($selectedRoles, $this->user);
-    }
-
-    /**
-     * Ověří požadovaný minimální věk.
-     *
-     * @throws SettingsItemNotFoundException
-     * @throws Throwable
-     */
-    public function validateRolesMinimumAge(MultiSelectBox $field): bool
-    {
-        $selectedRoles = $this->roleRepository->findRolesByIds($field->getValue());
-
-        return $this->validators->validateRolesMinimumAge($selectedRoles, $this->user);
-    }
-
-    /**
-     * Ověří požadovaný maximální věk.
-     *
-     * @throws SettingsItemNotFoundException
-     * @throws Throwable
-     */
-    public function validateRolesMaximumAge(MultiSelectBox $field): bool
-    {
-        $selectedRoles = $this->roleRepository->findRolesByIds($field->getValue());
-
-        return $this->validators->validateRolesMaximumAge($selectedRoles, $this->user);
     }
 }
